Add software version to RecordFileReader, update tests
