+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/ImporterIntegrationTest.java
@@ -21,7 +21,6 @@ import com.google.common.collect.Range;
 import com.hedera.mirror.common.config.CommonIntegrationTest;
 import com.hedera.mirror.common.converter.EntityIdConverter;
 import com.hedera.mirror.common.domain.entity.EntityId;
-import com.hedera.mirror.common.domain.transaction.RecordFile;
 import com.hedera.mirror.importer.ImporterIntegrationTest.Configuration;
 import com.hedera.mirror.importer.config.DateRangeCalculator;
 import com.hedera.mirror.importer.converter.JsonbToListConverter;
@@ -42,8 +41,6 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.reflect.FieldUtils;
@@ -141,11 +138,6 @@ public abstract class ImporterIntegrationTest extends CommonIntegrationTest {
         return findEntity(historyClass, ids, String.format("%s_history", table));
     }
 
-    protected void persistRecordFile(RecordFile recordFile) {
-        var migrationRecordFile = MigrationRecordFile.fromDomainRecordFile(recordFile);
-        perishRecordFile(migrationRecordFile);
-    }
-
     protected void reset() {
         super.reset();
         dateRangeCalculator.clear();
@@ -176,87 +168,6 @@ public abstract class ImporterIntegrationTest extends CommonIntegrationTest {
                 .collect(Collectors.joining(","));
 
         return !idColumns.isEmpty() ? idColumns : "id";
-    }
-
-    private void perishRecordFile(MigrationRecordFile migrationRecordFile) {
-        jdbcOperations.update(
-                """
-                    insert into record_file (bytes, consensus_end, consensus_start, count, digest_algorithm, file_hash, gas_used, hapi_version_major, hapi_version_minor, hapi_version_patch, hash, index, load_end, load_start, logs_bloom, name, node_id, prev_hash, sidecar_count, size, version)
-                    values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                    """,
-                migrationRecordFile.getBytes(),
-                migrationRecordFile.getConsensusEnd(),
-                migrationRecordFile.getConsensusStart(),
-                migrationRecordFile.getCount(),
-                migrationRecordFile.getDigestAlgorithm(),
-                migrationRecordFile.getFileHash(),
-                migrationRecordFile.getGasUsed(),
-                migrationRecordFile.getHapiVersionMajor(),
-                migrationRecordFile.getHapiVersionMinor(),
-                migrationRecordFile.getHapiVersionPatch(),
-                migrationRecordFile.getHash(),
-                migrationRecordFile.getIndex(),
-                migrationRecordFile.getLoadEnd(),
-                migrationRecordFile.getLoadStart(),
-                migrationRecordFile.getLogsBloom(),
-                migrationRecordFile.getName(),
-                migrationRecordFile.getNodeId(),
-                migrationRecordFile.getPreviousHash(),
-                migrationRecordFile.getSidecarCount(),
-                migrationRecordFile.getSize(),
-                migrationRecordFile.getVersion());
-    }
-
-    @Builder
-    @Data
-    protected static class MigrationRecordFile {
-        private byte[] bytes;
-        private long consensusStart;
-        private long consensusEnd;
-        private long count;
-        private int digestAlgorithm;
-        private String fileHash;
-        private long gasUsed;
-        private int hapiVersionMajor;
-        private int hapiVersionMinor;
-        private int hapiVersionPatch;
-        private String hash;
-        private long index;
-        private long loadStart;
-        private long loadEnd;
-        private byte[] logsBloom;
-        private String name;
-        private long nodeId;
-        private String previousHash;
-        private int sidecarCount;
-        private int size;
-        private int version;
-
-        public static MigrationRecordFile fromDomainRecordFile(RecordFile recordFile) {
-            return MigrationRecordFile.builder()
-                    .bytes(recordFile.getBytes())
-                    .consensusStart(recordFile.getConsensusStart())
-                    .consensusEnd(recordFile.getConsensusEnd())
-                    .count(recordFile.getCount())
-                    .digestAlgorithm(recordFile.getDigestAlgorithm().ordinal())
-                    .fileHash(recordFile.getFileHash())
-                    .gasUsed(recordFile.getGasUsed())
-                    .hapiVersionMajor(recordFile.getHapiVersionMajor())
-                    .hapiVersionMinor(recordFile.getHapiVersionMinor())
-                    .hapiVersionPatch(recordFile.getHapiVersionPatch())
-                    .hash(recordFile.getHash())
-                    .index(recordFile.getIndex())
-                    .loadStart(recordFile.getLoadStart())
-                    .loadEnd(recordFile.getLoadEnd())
-                    .logsBloom(recordFile.getLogsBloom())
-                    .name(recordFile.getName())
-                    .nodeId(recordFile.getNodeId())
-                    .previousHash(recordFile.getPreviousHash())
-                    .sidecarCount(recordFile.getSidecarCount())
-                    .size(recordFile.getSize())
-                    .version(recordFile.getVersion())
-                    .build();
-        }
     }
 
     @TestConfiguration(proxyBeanMethods = false)
