+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/reader/record/RecordFileReaderTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021-2024 Hedera Hashgraph, LLC
+ * Copyright (C) 2021-2025 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -105,6 +105,10 @@ abstract class RecordFileReaderTest {
                                     .boxed()
                                     .toList());
                     assertThat(transactionIndexes).doesNotHaveDuplicates().isSorted();
+
+                    assertThat(actual.getSoftwareVersionMajor()).isEqualTo(actual.getHapiVersionMajor());
+                    assertThat(actual.getSoftwareVersionMinor()).isEqualTo(actual.getHapiVersionMinor());
+                    assertThat(actual.getSoftwareVersionPatch()).isEqualTo(actual.getHapiVersionPatch());
                 });
     }
